\name{POMUtils-package}
\alias{POMUtils-package}
\alias{POMUtils}
\docType{package}
\title{
\packageTitle{POMUtils}
}
\description{
\packageDescription{POMUtils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{POMUtils}
\packageIndices{POMUtils}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{POMUtils}

Maintainer: \packageMaintainer{POMUtils}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
